app/utils/db_utils.py
"""
Ce module pourrait gérer la connexion à PostgreSQL/Redis quand ce sera prêt.
Pour l'instant, on laisse un squelette.
"""

import psycopg2
from psycopg2 import pool
from ..config import DATABASE_URL

# Pool de connexions (exemple)
try:
    connection_pool = psycopg2.pool.SimpleConnectionPool(
        1, 10, DATABASE_URL
    )
except:
    connection_pool = None
    print("Impossible de créer le pool de connexions.")

def get_connection():
    if connection_pool:
        return connection_pool.getconn()
    else:
        return None

def put_connection(conn):
    if connection_pool and conn:
        connection_pool.putconn(conn)
