requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
ortools==9.6.2534
psycopg2-binary==2.9.7
pydantic==1.10.7
pytest==7.4.0
requests==2.32.3
